Added roll, Added fan, Added Exit
Changing how the player can jump in water, added a roll when moving left or right to make the player look a bit more interactive. added a fan to push the player in a desired direction, works on windmills. added the escape root so the player has a mission objective.