Magnet alterations, and magnet skill added to metal material
Magnetise Metal blocks in the world whether they're in the back ground or foreground to effect the world outside of the players reach. Also the Magnet has been simplified so that the calculations aren't being done in the magnet blueprint but the actual cube objects themselves. I have removed the event tick from the magnet and put the force on a timer by event to reduce the amount on tick. Also the face has has a colour change and a death face. ICE NO LONGER DIES INSTANTLY IN WATER